See merge request polyflix-do/quizz!5

+++ src/main/modules/quizz/infrastructure/controllers/quizz.controller.ts
@@ -14,33 +14,51 @@ import {
 } from "../../application/dto/quizz.dto";
 import { QuizzParams } from "../adapters/params/quizz.param";
 import { QuizzService } from "../services/quizz.service";
+import { IsAdmin, MeId } from "@polyflix/x-utils";
 
 @Controller("quizzes")
 export class QuizzController {
   constructor(private readonly quizzService: QuizzService) {}
 
   @Get()
-  find(@Query() query: QuizzParams) {
-    return this.quizzService.find(query);
+  find(
+    @Query() query: QuizzParams,
+    @MeId() userId: string,
+    @IsAdmin() isAdmin: boolean
+  ) {
+    return this.quizzService.find(query, { userId, isAdmin });
   }
 
   @Get(":id")
-  findOne(@Param("id") id: string) {
-    return this.quizzService.findOne(id);
+  findOne(
+    @Param("id") id: string,
+    @MeId() userId: string,
+    @IsAdmin() isAdmin: boolean
+  ) {
+    return this.quizzService.findOne(id, { userId, isAdmin });
   }
 
   @Post()
-  create(@Body() createQuizzDto: CreateQuizzDTO) {
-    return this.quizzService.create(createQuizzDto, "");
+  create(@Body() createQuizzDto: CreateQuizzDTO, @MeId() userId: string) {
+    return this.quizzService.create(createQuizzDto, { userId });
   }
 
   @Put(":id")
-  update(@Param("id") id: string, @Body() updateQuizzDto: UpdateQuizzDTO) {
-    return this.quizzService.update(id, updateQuizzDto, "");
+  completeOne(
+    @Param("id") id: string,
+    @Body() updateQuizzDto: UpdateQuizzDTO,
+    @MeId() userId: string,
+    @IsAdmin() isAdmin: boolean
+  ) {
+    return this.quizzService.update(id, updateQuizzDto, { userId, isAdmin });
   }
 
   @Delete(":id")
-  deleteOne(@Param("id") id: string) {
-    return this.quizzService.delete(id, "");
+  deleteOne(
+    @Param("id") id: string,
+    @MeId() userId: string,
+    @IsAdmin() isAdmin: boolean
+  ) {
+    return this.quizzService.delete(id, { userId, isAdmin });
   }
 }
